Create year filter component
