+++ cinema_frontend/src/components/Filters/Filters.jsx
@@ -11,31 +11,15 @@ import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { ChevronDown } from "lucide-react";
 import GenreFilter from "./GenreFilter/GenreFilter";
+import YearFilter from "./YearFilter/YearFilter";
 
 const ageRatings = ["0+", "3+", "6+", "12+", "16+", "18+"];
 const ratingsRange = [0, 10]; 
 
 const Filters = () => {
-    const [selectedYear, setSelectedYear] = useState("");
     const [duration, setDuration] = useState([0, 240]);
     const [selectedAgeRating, setSelectedAgeRating] = useState("all");
     const [movieRating, setMovieRating] = useState([0, 10]); 
-
-    const currentYear = new Date().getFullYear();
-
-    const handleYearChange = (e) => {
-        const value = e.target.value;
-        setSelectedYear(value);
-    };
-
-    const handleYearBlur = () => {
-        const value = parseInt(selectedYear, 10);
-        if (value < 1930) {
-            setSelectedYear(1930);  
-        } else if (value > currentYear) {
-            setSelectedYear(currentYear); 
-        }
-    };
 
     return (
         <Popover>
@@ -50,21 +34,7 @@ const Filters = () => {
                     <GenreFilter />
 
                     {/* Year Filter */}
-                    <div>
-                        <Label htmlFor="year" className="font-semibold">
-                            Year (from 1930 to current year)
-                        </Label>
-                        <Input
-                            id="year"
-                            type="text"
-                            placeholder="Enter year"
-                            value={selectedYear}
-                            onChange={handleYearChange}
-                            onBlur={handleYearBlur} 
-                            min={1930} 
-                            max={currentYear}
-                        />
-                    </div>
+                    <YearFilter />
 
                     {/* Duration Filter */}
                     <div>
